Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,9 +2,9 @@
 
 int main() 
 {
-    float x,y,z;
+    int x,y,z;
     float a;
     scanf("%d %d %d",&x,&y,&z);
-    a=(x+y+z)/3;
+    a=(x+y+z) /3.0;
     printf("Average: %.2f",a);
 }
